requirements.txt
dj-database-url==0.5.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.8
django-filter==2.1.0
django-heroku==0.3.1
django-phonenumber-field==2.2.0
django-betterforms==1.2
Django==2.1.7
phonenumberslite==8.10.8
pytz==2018.9
rules==2.0.1
gunicorn==19.9.0
transliterate==1.10.2
whitenoise==4.1.2
psycopg2-binary==2.7.7
sqlparse==0.2.4
django-select2==7.0.3
